# openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: ADRO Control API
  version: 0.1.0
  description: Provider-independent Agentic Delivery & Release Orchestrator API.
  license:
    name: Apache License 2.0
    identifier: Apache-2.0
servers:
- url: "/"
security:
- bearerAuth: []
- cookieAuth: []
paths:
...
  "/api/v1/sessions":
    post:
      summary: Create an ADRO-owned durable harness session
      parameters:
      - "$ref": "#/components/parameters/IdempotencyKey"
      - "$ref": "#/components/parameters/RequestID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/HarnessSessionCreate"
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/JsonObject"
          headers:
            X-Request-ID:
              "$ref": "#/components/headers/RequestID"
            Idempotency-Replayed:
              "$ref": "#/components/headers/IdempotencyReplayed"
        default:
          description: Problem Details response
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/Problem"
            application/json:
              schema:
                "$ref": "#/components/schemas/JsonObject"
          headers:
            X-Request-ID:
              "$ref": "#/components/headers/RequestID"
  "/api/v1/sessions/{id}":
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: string
    get:
      summary: Read durable session and context status
      responses:
        '200':
          description: Session status
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/JsonObject"
          headers:
            X-Request-ID:
              "$ref": "#/components/headers/RequestID"
        default:
          description: Problem Details response
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/Problem"
            application/json:
              schema:
                "$ref": "#/components/schemas/JsonObject"
          headers:
            X-Request-ID:
              "$ref": "#/components/headers/RequestID"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    cookieAuth:
      type: apiKey
      in: cookie
      name: adro_session
  parameters:
...
    RequestID:
      name: X-Request-ID
      in: header
      schema:
        type: string
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Optional for POST, PUT, and PATCH. Scope is tenant + method + path.
        Reuse with the same query, content type, and body replays the first completed
        response; reuse with different input returns 409 idempotency_key_conflict.
      schema:
        type: string
        minLength: 1
        maxLength: 255
  schemas:
...
    HarnessSessionCreate:
      type: object
      required:
      - workspace_id
      properties:
        id:
          type: string
        workspace_id:
          type: string
        project_id:
          type: string
          description: Optional deterministic project-memory namespace shared by sessions.
        budget_tokens:
          type: integer
          minimum: 0
        auto_compaction:
          type: boolean
        compaction_threshold:
          type: number
          minimum: 0
          maximum: 1
        compaction_retain_tail:
          type: integer
          minimum: 0
...
    JsonObject:
      type: object
      description: Versioned JSON resource or page. Resource-specific fields remain
        backward-compatible within /api/v1.
      additionalProperties: true
    Problem:
      type: object
      required:
      - type
      - title
      - status
      - detail
      - error_code
      - request_id
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        error_code:
          type: string
        request_id:
          type: string
      additionalProperties: true
...
  headers:
    RequestID:
      description: Stable correlation identifier for this HTTP request.
      schema:
        type: string
    IdempotencyReplayed:
      description: True when the status, headers, and body were replayed from the
        first completed request.
      schema:
        type: string
        const: 'true'